pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ascii-cam"
version = "2.0.0"
description = "Terminal ASCII art converter with live camera support, video playback, audio visualization, and LLM context generation"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "ASCII Cam Developer"}
]
keywords = ["ascii", "art", "terminal", "camera", "webcam", "cli", "video", "audio", "llm", "ai"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Graphics :: Graphics Conversion",
    "Topic :: Multimedia :: Video",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "pillow>=10.0.0",
    "opencv-python>=4.8.0",
    "numpy>=1.24.0",
    "blessed>=1.20.0",
]

[project.optional-dependencies]
audio = ["pyaudio>=0.2.13"]
browser = ["selenium>=4.0.0", "playwright>=1.40.0"]
all = [
    "pyaudio>=0.2.13",
    "selenium>=4.0.0",
    "playwright>=1.40.0",
]

[project.scripts]
ascii-cam = "ascii_cam.cli:main"

[project.urls]
Homepage = "https://github.com/example/ascii-cam"

[tool.setuptools.packages.find]
where = ["."]
